Move SparseVector toDoQuery Finalization into Sparse Vector Field Mapping FieldType (#129020)
* move SparseVector toDoQuery final into field type

* [CI] Auto commit changes from spotless

* change method name to be specific to sparse vector

---------
